update dependencies, target pie, java 1.8 compat, move activities to a folder, fix some English, etc.
