+++ app/src/main/res/layout/fragment_select_mosaic.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
@@ -16,10 +15,10 @@
             <TextView
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:text="@string/mosaic_text"
+                android:layout_marginTop="30dp"
                 android:paddingLeft="30dp"
                 android:paddingRight="30dp"
-                android:layout_marginTop="30dp" />
+                android:text="@string/mosaic_text" />
 
             <Spinner
                 android:id="@+id/mosaicSpinner"
@@ -32,24 +31,24 @@
                 android:id="@+id/loopSwitch"
                 android:layout_width="match_parent"
                 android:layout_height="41dp"
-                android:text="@string/loop"
+                android:layout_marginTop="30dp"
                 android:paddingLeft="30dp"
                 android:paddingRight="30dp"
-                android:layout_marginTop="30dp"
+                android:text="@string/loop"
                 android:theme="@style/controlTheme" />
 
             <Button
                 android:id="@+id/viewButton"
+                style="@style/buttonTheme"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:text="@string/view_mosaic_image"
-                android:paddingLeft="30dp"
-                android:paddingRight="30dp"
-                android:layout_marginTop="50dp"
                 android:layout_marginLeft="20dp"
+                android:layout_marginTop="50dp"
                 android:layout_marginRight="20dp"
                 android:layout_marginBottom="20dp"
-                style="@style/buttonTheme" />
+                android:paddingLeft="30dp"
+                android:paddingRight="30dp"
+                android:text="@string/view_mosaic_image" />
 
         </LinearLayout>
     </ScrollView>
