+++ app/src/main/res/layout/fragment_select_nws.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
@@ -16,10 +15,10 @@
             <TextView
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:text="@string/location_text"
+                android:layout_marginTop="30dp"
                 android:paddingLeft="30dp"
                 android:paddingRight="30dp"
-                android:layout_marginTop="30dp" />
+                android:text="@string/location_text" />
 
             <Spinner
                 android:id="@+id/locationSpinner"
@@ -31,10 +30,10 @@
             <TextView
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:text="@string/type_text"
+                android:layout_marginTop="30dp"
                 android:paddingLeft="30dp"
                 android:paddingRight="30dp"
-                android:layout_marginTop="30dp" />
+                android:text="@string/type_text" />
 
             <Spinner
                 android:id="@+id/typeSpinner"
@@ -47,34 +46,34 @@
                 android:id="@+id/loopSwitch"
                 android:layout_width="match_parent"
                 android:layout_height="41dp"
-                android:text="@string/loop"
+                android:layout_marginTop="30dp"
                 android:paddingLeft="30dp"
                 android:paddingRight="30dp"
-                android:layout_marginTop="30dp"
+                android:text="@string/loop"
                 android:theme="@style/controlTheme" />
 
             <Switch
                 android:id="@+id/enhancedSwitch"
                 android:layout_width="match_parent"
                 android:layout_height="41dp"
-                android:text="@string/enhanced"
+                android:layout_marginTop="30dp"
                 android:paddingLeft="30dp"
                 android:paddingRight="30dp"
-                android:layout_marginTop="30dp"
+                android:text="@string/enhanced"
                 android:theme="@style/controlTheme" />
 
             <Button
                 android:id="@+id/viewButton"
+                style="@style/buttonTheme"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:text="@string/view_nws_image"
-                android:paddingLeft="30dp"
-                android:paddingRight="30dp"
-                android:layout_marginTop="50dp"
                 android:layout_marginLeft="20dp"
+                android:layout_marginTop="50dp"
                 android:layout_marginRight="20dp"
                 android:layout_marginBottom="20dp"
-                style="@style/buttonTheme" />
+                android:paddingLeft="30dp"
+                android:paddingRight="30dp"
+                android:text="@string/view_nws_image" />
 
         </LinearLayout>
     </ScrollView>
