+++ app/src/main/res/layout/fragment_select_wunderground.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
@@ -11,38 +10,38 @@
         <LinearLayout
             style="@style/layout_max_width"
             android:layout_height="wrap_content"
-            android:orientation="vertical"
             android:descendantFocusability="beforeDescendants"
             android:focusable="true"
-            android:focusableInTouchMode="true">
+            android:focusableInTouchMode="true"
+            android:orientation="vertical">
 
             <TextView
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:text="@string/location_text"
+                android:layout_marginTop="30dp"
                 android:paddingLeft="30dp"
                 android:paddingRight="30dp"
-                android:layout_marginTop="30dp" />
+                android:text="@string/location_text" />
 
             <view
-                class="com.danhasting.radar.views.EditTextDefocus"
                 android:id="@+id/wunderground_location"
+                class="com.danhasting.radar.views.EditTextDefocus"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
+                android:layout_marginLeft="30dp"
+                android:layout_marginRight="30dp"
                 android:ems="10"
                 android:hint="@string/location_placeholder"
                 android:inputType="textCapWords"
-                android:layout_marginLeft="30dp"
-                android:layout_marginRight="30dp"
                 android:selectAllOnFocus="true" />
 
             <TextView
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:text="@string/type_text"
+                android:layout_marginTop="30dp"
                 android:paddingLeft="30dp"
                 android:paddingRight="30dp"
-                android:layout_marginTop="30dp" />
+                android:text="@string/type_text" />
 
             <Spinner
                 android:id="@+id/wundergroundTypeSpinner"
@@ -55,36 +54,36 @@
                 android:id="@+id/loopSwitch"
                 android:layout_width="match_parent"
                 android:layout_height="41dp"
-                android:text="@string/loop"
+                android:layout_marginTop="30dp"
                 android:paddingLeft="30dp"
                 android:paddingRight="30dp"
-                android:layout_marginTop="30dp"
+                android:text="@string/loop"
                 android:theme="@style/controlTheme" />
 
             <TextView
                 android:id="@+id/radiusText"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:text="@string/radius_text"
+                android:layout_marginTop="30dp"
                 android:paddingLeft="30dp"
                 android:paddingRight="30dp"
-                android:layout_marginTop="30dp" />
+                android:text="@string/radius_text" />
 
             <LinearLayout
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:orientation="horizontal"
-                android:layout_marginTop="10dp">
+                android:layout_marginTop="10dp"
+                android:orientation="horizontal">
 
                 <SeekBar
                     android:id="@+id/radiusBar"
                     android:layout_width="0dp"
-                    android:layout_weight="1"
                     android:layout_height="30dp"
-                    android:layout_marginLeft="30dp"
                     android:layout_marginStart="30dp"
-                    android:layout_marginRight="10dp"
+                    android:layout_marginLeft="30dp"
                     android:layout_marginEnd="10dp"
+                    android:layout_marginRight="10dp"
+                    android:layout_weight="1"
                     android:theme="@style/controlTheme" />
 
                 <TextView
@@ -97,16 +96,16 @@
 
             <Button
                 android:id="@+id/viewButton"
+                style="@style/buttonTheme"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_marginBottom="20dp"
                 android:layout_marginLeft="20dp"
-                android:layout_marginRight="20dp"
                 android:layout_marginTop="50dp"
+                android:layout_marginRight="20dp"
+                android:layout_marginBottom="20dp"
                 android:paddingLeft="30dp"
                 android:paddingRight="30dp"
-                android:text="@string/view_wunderground_image"
-                style="@style/buttonTheme" />
+                android:text="@string/view_wunderground_image" />
 
         </LinearLayout>
     </ScrollView>
