+++ app/src/main/res/layout/nav_header.xml
@@ -2,9 +2,9 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="?android:attr/actionBarSize"
+    android:gravity="center_vertical"
     android:paddingLeft="16dp"
     android:paddingRight="16dp"
-    android:gravity="center_vertical"
     android:theme="@style/NavHeader">
 
     <TextView
@@ -12,19 +12,19 @@
         android:layout_height="28dp"
         android:layout_marginEnd="20dp"
         android:layout_marginRight="20dp"
-        android:text="@string/header_title"
         android:layout_toStartOf="@id/nav_settings"
         android:layout_toLeftOf="@id/nav_settings"
+        android:text="@string/header_title"
         android:textAppearance="@style/TextAppearance.AppCompat.Title" />
 
     <Button
         android:id="@+id/nav_settings"
         android:layout_width="28dp"
         android:layout_height="28dp"
-        android:layout_marginEnd="40dp"
-        android:layout_marginRight="40dp"
         android:layout_alignParentEnd="true"
         android:layout_alignParentRight="true"
+        android:layout_marginEnd="40dp"
+        android:layout_marginRight="40dp"
         android:background="@drawable/ic_settings" />
 
     <Button
