+++ app/src/main/res/menu/drawer_view.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android" >
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
     <group android:checkableBehavior="single">
 
         <item
@@ -19,8 +19,8 @@
 
         <item
             android:id="@+id/nav_favorites"
-            android:titleCondensed="@string/nav_favorites"
-            android:title="@string/nav_favorites">
+            android:title="@string/nav_favorites"
+            android:titleCondensed="@string/nav_favorites">
             <menu />
         </item>
     </group>
